+++ fabric/src/main/java/io/github/sakurawald/fuji/core/command/executor/CommandExecutor.java
@@ -36,7 +36,7 @@ public class CommandExecutor {
 
         /* Expand the command. */
         command = context.expandCommand(command);
-        LogUtil.debug("Execute the command: command = `{}`, context = {}", command, context);
+        LogUtil.debug("Command Executor -> Execute the command: command = `{}`, context = {}", command, context);
 
         try {
             return Objects