+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_advice/structure/CommandAdviceEntry.java
@@ -61,7 +61,4 @@ public class CommandAdviceEntry {
         """)
     List<String> commands;
 
-    public boolean isCancellableAdviceType() {
-        return this.getAdviceType().isCancellable();
-    }
 }