+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_advice/structure/CommandAdviceType.java
@@ -1,23 +1,28 @@
 package io.github.sakurawald.fuji.module.initializer.command_advice.structure;
 
 import io.github.sakurawald.fuji.core.auxiliary.minecraft.CommandHelper;
+import io.github.sakurawald.fuji.core.document.annotation.ForDeveloper;
 import lombok.Getter;
 
 @Getter
 public enum CommandAdviceType {
-    BEFORE_EXECUTING(false, CommandHelper.Return.SUCCESS),
-    AFTER_EXECUTING(false, CommandHelper.Return.SUCCESS),
-    CANCEL_WITH_SUCCESS(true, CommandHelper.Return.SUCCESS),
-    CANCEL_WITH_FAILURE(true, CommandHelper.Return.FAILURE),
-    CANCEL_IF_ANY_SUCCESS(true, CommandHelper.Return.FAILURE),
-    CANCEL_IF_ALL_SUCCESS(true, CommandHelper.Return.FAILURE);
+    BEFORE_EXECUTING(false, false, CommandHelper.Return.SUCCESS),
+    AFTER_EXECUTING(false, false, CommandHelper.Return.SUCCESS),
+    CANCEL_WITH_SUCCESS(true, false, CommandHelper.Return.SUCCESS),
+    CANCEL_WITH_FAILURE(true, false, CommandHelper.Return.FAILURE),
+    CANCEL_IF_ANY_SUCCESS(true, false, CommandHelper.Return.FAILURE),
+    CANCEL_IF_ALL_SUCCESS(true, false, CommandHelper.Return.FAILURE),
+    ON_CANCELLED(false, true, CommandHelper.Return.SUCCESS),;
 
-    final boolean isCancellable;
+    @ForDeveloper("A canceller advice will be accepted by BEFORE_EXECUTING type.")
+    final boolean isCanceller;
+    @ForDeveloper("A monitor advice will be accepted by any type.")
+    final boolean isMonitor;
     final int alternativeReturnValue;
 
-    CommandAdviceType(boolean isCancellable, int alternativeReturnValue) {
+    CommandAdviceType(boolean isCanceller, boolean isMonitor, int alternativeReturnValue) {
+        this.isCanceller = isCanceller;
+        this.isMonitor = isMonitor;
         this.alternativeReturnValue = alternativeReturnValue;
-        this.isCancellable = isCancellable;
     }
-
 }